+++ dflib-echarts/src/main/java/org/dflib/echarts/SeriesModelMaker.java
@@ -87,7 +87,7 @@ class SeriesModelMaker {
                 null,
                 null,
                 null,
-                null,
+                sb.seriesOpts.symbolSize,
                 sb.seriesOpts.xAxisIndex,
                 sb.seriesOpts.yAxisIndex,
                 null,